Init Essentia melspectrogram extractor on demand
Lighting fails to parallelize the network due to the fact that multi-threading is not supported in
Essentia.
As a simple solution, the Essentia extractor is initialized on demand, that is, when a 1D signal
(audio assumed) is forwarded to the network. This is intended to happed outside the training loop
exclusively, since the training phase is intended to operate with mel-specs, thus, avoiding the
crash between Essentia and Lighting.

Note: Training directly on waveforms would require finding a parallelizable alternative to Essentia.