Make MAEST installable via pyproject.toml
It is configured to install the `models/` package only under the name
`maest`.
The `maest` function is renamed to `get_maest` to avoid confusion with
the package name.
Finally, `get_maest` is exposed in the `models` package for convenience.

This way, when the package is installed, the model can be imported as:
`from maest impot get_maest`
Otherwise, it can be imported as:
`from models import get_maest`